+++ src/Introspecter.tsx
@@ -1,69 +1,10 @@
 import React, { useContext, useEffect, useMemo } from 'react';
 import { useLogoutIfAccessDenied } from 'react-admin';
+
 import SchemaAnalyzerContext from './SchemaAnalyzerContext.js';
 import useIntrospect from './useIntrospect.js';
-import type {
-  IntrospecterProps,
-  ResourcesIntrospecterProps,
-  SchemaAnalyzer,
-} from './types.js';
-
-const ResourcesIntrospecter = ({
-  component: Component,
-  schemaAnalyzer,
-  includeDeprecated,
-  resource,
-  resources,
-  loading,
-  error,
-  ...rest
-}: ResourcesIntrospecterProps) => {
-  if (loading) {
-    return null;
-  }
-
-  if (error) {
-    if (process.env.NODE_ENV === 'production') {
-      // eslint-disable-next-line no-console
-      console.error(error);
-    }
-
-    throw new Error('API schema is not readable');
-  }
-
-  const schema = resources.find((r) => r.name === resource);
-
-  if (!schema?.fields || !schema?.readableFields || !schema?.writableFields) {
-    if (process.env.NODE_ENV === 'production') {
-      // eslint-disable-next-line no-console
-      console.error(`Resource ${resource} not present inside API description`);
-    }
-
-    throw new Error(`Resource ${resource} not present inside API description`);
-  }
-
-  const fields = includeDeprecated
-    ? schema.fields
-    : schema.fields.filter(({ deprecated }) => !deprecated);
-  const readableFields = includeDeprecated
-    ? schema.readableFields
-    : schema.readableFields.filter(({ deprecated }) => !deprecated);
-  const writableFields = includeDeprecated
-    ? schema.writableFields
-    : schema.writableFields.filter(({ deprecated }) => !deprecated);
-
-  return (
-    <Component
-      schemaAnalyzer={schemaAnalyzer}
-      resource={resource}
-      schema={schema}
-      fields={fields}
-      readableFields={readableFields}
-      writableFields={writableFields}
-      {...rest}
-    />
-  );
-};
+import type { IntrospecterProps, SchemaAnalyzer } from './types.js';
+import ResourcesIntrospecter from './ResourcesIntrospecter.js';
 
 const Introspecter = ({
   component,
@@ -106,7 +47,8 @@ const Introspecter = ({
       },
     });
   }, [schemaAnalyzer, logoutIfAccessDenied]);
-  const { refetch, data, isLoading, isIdle, error } = useIntrospect();
+
+  const { refetch, data, isPending, error } = useIntrospect();
   const resources = data ? data.data.resources : null;
 
   useEffect(() => {
@@ -126,7 +68,7 @@ const Introspecter = ({
       includeDeprecated={includeDeprecated}
       resource={resource}
       resources={resources ?? []}
-      loading={isLoading || isIdle}
+      loading={isPending}
       error={error}
       {...rest}
     />
